# apps/agent/src/api/openapi.yaml
openapi: 3.1.0
info:
  title: LeadSwap API
  version: 1.0.0
  description: |
    AI-powered B2B lead validation API.
    
    LeadSwap validates your leads against your Ideal Customer Profile (ICP)
    before you buy them, saving credits and improving conversion rates.
    
    ## How it works
    1. Define your ICP in natural language
    2. Upload a CSV of leads
    3. Run validation
    4. Get scored results with actionable insights

servers:
  - url: https://your-domain.com
    description: Production server
  - url: http://localhost:3000
    description: Development server

paths:
  /api/icp:
    post:
      operationId: defineICP
      summary: Define your Ideal Customer Profile
      description: |
        Define your ICP using natural language. Examples:
        - "SaaS startups, 50-200 employees, France, VP Sales"
        - "FinTech companies in Europe, Series B+, Head of Growth"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - userId
                - input
              properties:
                userId:
                  type: string
                  description: Unique identifier for the user
                  example: "user123"
                input:
                  type: string
                  description: Natural language ICP description
                  example: "SaaS startups, 50-200 employees, France, VP Sales"
      responses:
        '200':
          description: ICP defined successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ICPResponse'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /api/icp/{userId}:
    get:
      operationId: getICP
      summary: Get the ICP for a user
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: ICP retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ICPResponse'
        '404':
          description: ICP not found

  /api/upload:
    post:
      operationId: uploadLeads
      summary: Upload a CSV of leads
      description: |
        Upload leads from a CSV file. The system will auto-detect columns:
        - email (required)
        - name / first_name + last_name
        - company
        - title
        - linkedin_url
        - website
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - userId
              properties:
                userId:
                  type: string
                  description: User identifier
                csvContent:
                  type: string
                  description: Raw CSV content as string
                csvBase64:
                  type: string
                  description: Base64 encoded CSV content
      responses:
        '200':
          description: CSV parsed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UploadResponse'
        '400':
          description: Parse error

  /api/validate:
    post:
      operationId: validateLeads
      summary: Validate leads against ICP
      description: |
        Run validation on uploaded leads. Each lead is scored 0-100 and
        assigned to a tier:
        - Tier 1 (80-100): Hot leads, ready to contact
        - Tier 2 (50-79): Warm leads, worth pursuing
        - Rejected (0-49): Don't buy these
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - userId
              properties:
                userId:
                  type: string
                  description: User identifier
      responses:
        '200':
          description: Validation complete
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationResponse'
        '400':
          description: Missing ICP or leads

  /api/validate/results/{id}:
    get:
      operationId: getResults
      summary: Get validation results
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - name: tier
          in: query
          schema:
            type: string
            enum: [tier1, tier2, rejected]
          description: Filter by tier
      responses:
        '200':
          description: Results retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResultsResponse'

  /api/validate/export/{id}:
    get:
      operationId: exportResults
      summary: Export results as CSV
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - name: tier
          in: query
          schema:
            type: string
            enum: [tier1, tier2, qualified, all]
          description: Which leads to export
      responses:
        '200':
          description: CSV file
          content:
            text/csv:
              schema:
                type: string

components:
  schemas:
    ICPResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        icp:
          $ref: '#/components/schemas/ICP'
        displayText:
          type: string
          description: Human-readable ICP summary
        confidence:
          type: number
          description: Parsing confidence 0-1
        warnings:
          type: array
          items:
            type: string

    ICP:
      type: object
      properties:
        userId:
          type: string
        industries:
          type: array
          items:
            type: string
        companySizes:
          type: object
          properties:
            min:
              type: integer
            max:
              type: integer
        geoTargets:
          type: array
          items:
            type: string
        titles:
          type: array
          items:
            type: string
        keywords:
          type: array
          items:
            type: string

    UploadResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        displayText:
          type: string
        stats:
          type: object
          properties:
            totalRows:
              type: integer
            validRows:
              type: integer
            invalidRows:
              type: integer
        columnMapping:
          type: object
        warnings:
          type: array
          items:
            type: string

    ValidationResponse:
      type: object
      properties:
        success:
          type: boolean
        resultId:
          type: string
        displayText:
          type: string
          description: Formatted results summary
        summary:
          type: object
          properties:
            tier1:
              type: object
              properties:
                count:
                  type: integer
                percentage:
                  type: integer
            tier2:
              type: object
              properties:
                count:
                  type: integer
                percentage:
                  type: integer
            rejected:
              type: object
              properties:
                count:
                  type: integer
                percentage:
                  type: integer
            avgScore:
              type: integer
            creditsSaved:
              type: integer
              description: Estimated credits saved in euros
        patterns:
          type: array
          items:
            $ref: '#/components/schemas/Pattern'
        recommendations:
          type: array
          items:
            type: string

    ResultsResponse:
      type: object
      properties:
        success:
          type: boolean
        resultId:
          type: string
        tier:
          type: string
        leadCount:
          type: integer
        leads:
          type: array
          items:
            $ref: '#/components/schemas/ScoredLead'

    ScoredLead:
      type: object
      properties:
        id:
          type: string
        email:
          type: string
        name:
          type: string
        company:
          type: string
        title:
          type: string
        score:
          type: integer
          description: 0-100 score
        tier:
          type: string
          enum: [tier1, tier2, rejected]
        reasoning:
          type: array
          items:
            type: string
        intentSignals:
          type: array
          items:
            type: object
            properties:
              type:
                type: string
              description:
                type: string
              emoji:
                type: string
              boost:
                type: integer

    Pattern:
      type: object
      properties:
        reason:
          type: string
        count:
          type: integer
        percentage:
          type: integer
        recommendation:
          type: string

    Error:
      type: object
      properties:
        error:
          type: string
        message:
          type: string